[feat] 존재하지 않는 이미지 삭제시 FileNotExistsException 예외 처리 구현

+++ src/main/java/com/haejwo/tripcometrue/global/s3/service/S3Service.java
@@ -3,6 +3,7 @@ package com.haejwo.tripcometrue.global.s3.service;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.haejwo.tripcometrue.global.s3.exception.FileEmptyException;
+import com.haejwo.tripcometrue.global.s3.exception.FileNotExistsException;
 import com.haejwo.tripcometrue.global.s3.exception.FileUploadFailException;
 import com.haejwo.tripcometrue.global.s3.response.S3UploadResponseDto;
 import lombok.RequiredArgsConstructor;
@@ -24,8 +25,10 @@ public class S3Service {
     @Value("${cloud.aws.s3.bucket}")
     private String bucketName;
 
+    // TODO: 1/6/24 멀티파트 파일을 일반 파일로 전환??
     // TODO: 1/6/24 파일 용량 제한 걸기
-    // TODO: 1/6/24 저장 디렉토리 분리하기
+    // TODO: 1/6/24 저장 디렉토리 분리하기??
+    // TODO: 1/6/24 개발 버킷과 배포 버킷을 구분하기
     public S3UploadResponseDto saveImage(MultipartFile multipartFile) {
         validateFileExists(multipartFile);
         String filename = generateFilename(multipartFile);
@@ -38,6 +41,10 @@ public class S3Service {
         return new S3UploadResponseDto(amazonS3.getUrl(bucketName, filename).toString());
     }
 
+    private String generateFilename(MultipartFile multipartFile) {
+        return UUID.randomUUID() + "_" + multipartFile.getOriginalFilename();
+    }
+
     private ObjectMetadata getObjectMetadata(MultipartFile multipartFile) {
         ObjectMetadata metadata = new ObjectMetadata();
         metadata.setContentLength(multipartFile.getSize());
@@ -45,13 +52,12 @@ public class S3Service {
         return metadata;
     }
 
-    private String generateFilename(MultipartFile multipartFile) {
-        return UUID.randomUUID() + "_" + multipartFile.getOriginalFilename();
-    }
-
-    // TODO: 1/6/24 파일이 존재하지 않을 경우 예외 발생 추가
     public void removeImage(String url) {
         String filename = getFilename(url);
+        if (!amazonS3.doesObjectExist(bucketName, filename)) {
+            throw new FileNotExistsException();
+        }
+
         amazonS3.deleteObject(bucketName, filename);
     }
